Split contract and component code to different files
