+++ crates/cairo-lang-starknet/src/plugin/mod.rs
@@ -10,12 +10,11 @@ use cairo_lang_syntax::node::helpers::QueryAttrs;
 use consts::*;
 
 pub mod aux_data;
-mod contract;
 mod dispatcher;
 mod entry_point;
 pub mod events;
-mod generation_data;
 mod includable;
+mod starknet_module;
 mod storage;
 mod storage_access;
 mod utils;
@@ -24,8 +23,8 @@ use dispatcher::handle_trait;
 use events::derive_event_needed;
 use storage_access::derive_storage_access_needed;
 
-use self::contract::{handle_module, handle_module_by_storage};
 use self::includable::handle_includable;
+use self::starknet_module::{handle_module, handle_module_by_storage};
 
 #[derive(Debug, Default)]
 #[non_exhaustive]
