+++ crates/cairo-lang-starknet/src/plugin/plugin_test_data/components/component
@@ -159,7 +159,6 @@ use starknet::event::EventEmitter;
     }
 
 
-
 impls:
 
 impl EventDrop of Drop::<Event>;
