+++ crates/cairo-lang-starknet/src/plugin/plugin_test_data/contracts/contract
@@ -265,7 +265,6 @@ fn __wrapper_test_serde(mut data: Span::<felt252>) -> Span::<felt252> {
     }
 
 
-
 impls:
 
 impl EventDrop of Drop::<Event>;
