+++ crates/cairo-lang-starknet/src/plugin/plugin_test_data/contracts/included_impl
@@ -218,7 +218,6 @@ impl ContractStateOutsideImpl of
     }
 
 
-
 impls:
 
 impl EventDrop of Drop::<Event>;
