+++ crates/cairo-lang-starknet/src/plugin/plugin_test_data/contracts/l1_handler
@@ -131,7 +131,6 @@ fn __wrapper_good_l1_handler_ignored(mut data: Span::<felt252>) -> Span::<felt25
     }
 
 
-
 impls:
 
 impl EventDrop of Drop::<Event>;
