+++ crates/cairo-lang-starknet/src/plugin/plugin_test_data/contracts/raw_output
@@ -121,7 +121,6 @@ fn __wrapper_test_raw_output_with_spaces(mut data: Span::<felt252>) -> Span::<fe
     }
 
 
-
 impls:
 
 impl EventDrop of Drop::<Event>;
