+++ crates/cairo-lang-starknet/src/plugin/plugin_test_data/contracts/storage
@@ -484,7 +484,6 @@ const TEST_CLASS_HASH: felt252 = 0xdf00ddc8940a9ebc2f8665e989d2652203dabc0967e7b
     }
 
 
-
 impls:
 
 impl EventDrop of Drop::<Event>;
