+++ crates/cairo-lang-starknet/src/plugin/starknet_module/generation_data.rs
@@ -1,6 +1,7 @@
 use cairo_lang_defs::patcher::RewriteNode;
 
-use super::contract::{ComponentSpecificGenerationData, ContractSpecificGenerationData};
+use super::component::ComponentSpecificGenerationData;
+use super::contract::ContractSpecificGenerationData;
 
 /// The data for generating the code of a contract.
 #[derive(Default)]
